+++ lib/test_prof/event_prof/profiler.rb
@@ -3,7 +3,8 @@
 module TestProf
   module EventProf
     class Profiler # :nodoc:
-      attr_reader :event, :total_count, :total_time, :rank_by, :top_count, :per_example
+      attr_reader :event, :total_count, :total_time, :rank_by, :top_count, :per_example,
+                  :time, :count, :example_time, :example_count
 
       alias per_example? per_example